Separate tx new commands
